email bot
waits to receive an email, gathers the email addresses from "from" and "to," and sends an email response back. Currently it sends the response email to everyone originally listed in the "to" section, but it can have one line added to it to include a response to the original sender. Only works with gmail since it utilizes their api.